2022/09/29 댓글, 대댓글 등록, 삭제 제어, 코드 정리, 목록 버튼 경로 재설정

+++ src/main/java/com/adnstyle/choicafe/service/GhReplyService.java
@@ -1,16 +1,20 @@
 package com.adnstyle.choicafe.service;
 
 
+import com.adnstyle.choicafe.common.SessionMember;
 import com.adnstyle.choicafe.domain.GhReply;
 import com.adnstyle.choicafe.repository.GhReplyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class GhReplyService {
+
+    private final HttpSession httpSession;
 
     private final GhReplyRepository ghReplyRepository;
 
@@ -38,6 +42,13 @@ public class GhReplyService {
 
     }
     public String deleteRely(GhReply ghReply) {
+
+        SessionMember ghMember = (SessionMember) httpSession.getAttribute("user");
+
+        if (!ghReply.getMemberSeq().equals(ghMember.getSeq()) && ghMember.getRole().equals("ROLE_ADMIN")){
+            return "fail";
+        }
+
         int chk = ghReplyRepository.deleteReply(ghReply);
 
         if (chk > 0) {
